refactor: dedupe HTTP middleware and tidy legacy docs (#107)
Extract serverhttp.HandlerMiddleware for gatewayserver/https, warn on legacy grpc passthrough mode, and update p2p/legacy documentation.
