+++ servers/grpcs/doc.go
@@ -1,0 +1,5 @@
+// Package grpcs is a deprecated alias for gatewayserver.
+//
+// Deprecated: import github.com/pubgo/lava/v2/servers/gatewayserver instead.
+// This package will be removed in v3.
+package grpcs
